requirements.txt
beautifulsoup4
requests
python-dotenv
Pillow
anthropic
fastapi
uvicorn
pydantic
